requirements_lock.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements_lock.txt requirements.txt
#
cnunciato-buildkite-sdk==0.0.49
    # via -r requirements.txt
pyyaml==6.0.2
    # via cnunciato-buildkite-sdk
